reorganized files according to new classes
