+++ tdarr-noding/src/configuration_parsing/constant_creator.py
