+++ tdarr-noding/src/configuration_parsing/node_class.py
@@ -1,0 +1,2 @@
+class Node:
+    def __init__(self, node_name, node_id_inner_dictionary):
